userService/requirements.txt

Django==4.1.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2>=2.8
#pandas==1.4.2
pyjwt==2.6.0
requests==2.28.2
pika==1.3.1






